pyproject.toml
[project]
name = "discord-conversation-synthesizer"
version = "0.1.0"
description = "Transform Discord conversations into structured Obsidian knowledge"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

authors = [
    {name = "Valter", email = "valterallen97@gmail.com"}
]

dependencies = [
    "discord.py>=2.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
synthesizer = "src.__main__:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
